requirements.txt
flask_cors
flask_sqlalchemy
flask_migrate
requests
requests-cache
